Move factory to reader to avoid circular import, adjust metadata to have proper parent child mapping and adjust all needed code to accomodate
